development
minor improvements to existing features, some bug fixes, implemented some features that apply to selected elements,

multiple selection via shift+left click,

use context menu on element, to change values for selection

movement of multiple elements is buggy, the initial coordinates are not mapped correctly

+++ Commands/EditEBoardParametersCommand.cs
@@ -28,8 +28,8 @@ namespace EBoard.Commands
             {
                 _EBoardBrowserViewModel.SelectedEBoard.EBoardName = _EBoardBrowserViewModel.EBoardName;
                 _EBoardBrowserViewModel.SelectedEBoard.EBoardDepth = _EBoardBrowserViewModel.EBoardDepth;
-                _EBoardBrowserViewModel.SelectedEBoard.Width = _EBoardBrowserViewModel.NewEBoardWidth;
-                _EBoardBrowserViewModel.SelectedEBoard.Height = _EBoardBrowserViewModel.NewEBoardHeight;
+                _EBoardBrowserViewModel.SelectedEBoard.Width = (int)_EBoardBrowserViewModel.NewEBoardWidth;
+                _EBoardBrowserViewModel.SelectedEBoard.Height = (int)_EBoardBrowserViewModel.NewEBoardHeight;
             }
 
         }
